Fix playback of files with unicode filenames
Problem showed up on OSX where a file from an NFS mounted directory
would not open because unicode filename included an accent -- Françoise.
However, if the file was stored in a local directory it would play.
The fix is to convert the unicode filename to NFD normalization form
prior to converting to opening the file.

git-svn-id: https://svn.code.sf.net/p/apprenticevideo/code/trunk@452 28900df4-563f-46a4-9db8-70c0268b8f86