Stop using GL_APPLE_client_storage to fix out of order frame rendering on the Mac
GL_APPLE_client_storage avoids copying the texture data.  However,
this creates a race condition between the graphics card painting from
an about-to-be reused texture buffer and ffmpeg decoder that will
reuse the texture buffer.  One fix is to double buffer the
frames in the Canvas, but it is simpler to not use
GL_APPLE_client_storage in the first place, since there is no apparent
benefit.

git-svn-id: https://svn.code.sf.net/p/apprenticevideo/code/trunk@528 28900df4-563f-46a4-9db8-70c0268b8f86