Split the_soft_point_t into the_soft_point_t and the_sticky_point_t
In revision 36 the_soft_point_t behavior was changed to modify the
reference location during regeneration.  This is not the originally
intended behavior of the soft point, although it is beneficial when
editing a curve that supports other curves.

This checkin restores the original behavior of the_soft_point_t, and
adds the_sticky_point_t which calculates the point position as the
closest position on the reference geometry.